refactor table validation so index validation can rely on schemaexpr package
