+++ pkg/sql/sqlbase/partition.go
@@ -17,7 +17,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/util/encoding"
-	"github.com/cockroachdb/cockroach/pkg/util/interval"
 	"github.com/cockroachdb/errors"
 )
 
@@ -185,20 +184,4 @@ func DecodePartitionTuple(
 	}
 
 	return t, key, nil
-}
-
-type partitionInterval struct {
-	name  string
-	start roachpb.Key
-	end   roachpb.Key
-}
-
-var _ interval.Interface = partitionInterval{}
-
-// ID is part of `interval.Interface` but unused in validatePartitioningDescriptor.
-func (ps partitionInterval) ID() uintptr { return 0 }
-
-// Range is part of `interval.Interface`.
-func (ps partitionInterval) Range() interval.Range {
-	return interval.Range{Start: []byte(ps.start), End: []byte(ps.end)}
 }
